Extracted some boiler code into seperate files. Removed .env file from .gitignore and added logs and comments for better readability.
