+++ .env
@@ -1,0 +1,6 @@
+OPENAI_API_KEY=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
+
+# Phone number format: 972541234567 
+# 972 Area Code(for Israel).
+# For this phone number: 054-1234567 
+CHAT_BOT_PHONE_NUMBER=
