+++ .gitignore
@@ -1,5 +1,4 @@
 /node_modules
-/.env
 /.wwebjs_cache
 /.wwebjs_auth
 /package-lock.json
